backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
openai==1.55.0
requests==2.32.3
pydantic==2.10.0
python-dotenv==1.0.1